Release 11.15.04
git-svn-id: http://svn.code.sf.net/p/netpbm/code/advanced@5276 9d0c8265-081b-0410-96cb-a4ca84ce46f8

+++ converter/pbm/mdatopbm.c
@@ -279,8 +279,8 @@ main(int argc, const char **argv) {
 
     {
         short yy;
-        pm_readlittleshort(ifP, &yy); nInRows = yy;
-        pm_readlittleshort(ifP, &yy); nInCols = yy;
+        pm_readlittleshort(ifP, &yy); nInRows = (unsigned short)yy;
+        pm_readlittleshort(ifP, &yy); nInCols = (unsigned short)yy;
     }
 
     nOutCols = 8 * nInCols;
